Medical_ai_agent/requirements.txt
# ╔══════════════════════════════════════════════════════════╗
# ║        MEDICAL AI AGENT - requirements.txt               ║
# ║  Install all: pip install -r requirements.txt            ║
# ╚══════════════════════════════════════════════════════════╝

# ── Core Framework ──────────────────────────────────────────
streamlit>=1.35.0
python-dotenv>=1.0.0

# ── LLM / AI ────────────────────────────────────────────────
openai>=1.40.0
anthropic>=0.30.0                # Optional: Claude support
langchain>=0.2.0
langchain-openai>=0.1.10
langchain-community>=0.2.0
langchain-pinecone>=0.1.3
crewai>=0.51.0                   # Multi-agent orchestration

# ── Vector Database ──────────────────────────────────────────
pinecone-client>=4.0.0
pinecone>=4.0.0

# ── Graph Database ───────────────────────────────────────────
neo4j>=5.20.0
graphrag>=0.3.0                  # Microsoft GraphRAG

# ── Document Parsing ─────────────────────────────────────────
pdfplumber>=0.11.0
pypdf>=4.2.0
pytesseract>=0.3.10
Pillow>=10.3.0
unstructured>=0.14.0
unstructured[pdf]>=0.14.0

# ── NLP / Text Processing ────────────────────────────────────
spacy>=3.7.0
# python -m spacy download en_core_web_sm
# python -m spacy download en_ner_bc5cdr_md  # Biomedical NER (optional)
transformers>=4.40.0             # HuggingFace models
sentence-transformers>=3.0.0    # Semantic embeddings
rank_bm25>=0.2.2                # BM25 sparse retrieval

# ── Medical NLP (optional, powerful) ────────────────────────
# scispacy>=0.5.4                # Scientific text NLP
# medspacy>=1.1.0                # Clinical NLP

# ── Voice Input (optional) ───────────────────────────────────
openai-whisper>=20231117        # Audio transcription

# ── Alerting (optional) ──────────────────────────────────────
twilio>=9.0.0                   # SMS/email alerts

# ── Translation (optional) ───────────────────────────────────
deepl>=1.18.0                   # Multi-language output

# ── Evaluation ───────────────────────────────────────────────
rouge-score>=0.1.2              # ROUGE evaluation
nltk>=3.8.1                     # BLEU and NLP utilities
# nltk.download('punkt')

# ── Export / Reports ────────────────────────────────────────
reportlab>=4.2.0                # PDF report generation
fpdf2>=2.7.9                    # Alternative PDF library
python-docx>=1.1.0              # DOCX export

# ── Data Processing ─────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0

# ── Monitoring ───────────────────────────────────────────────
langsmith>=0.1.75               # LangChain tracing

# ── Bioinformatics (optional) ────────────────────────────────
biopython>=1.83                 # PubMed Entrez API access

# ── Utilities ────────────────────────────────────────────────
requests>=2.32.0
httpx>=0.27.0
tqdm>=4.66.0
pydantic>=2.7.0
loguru>=0.7.2
